+++ R/zzz.r
@@ -1,12 +1,13 @@
+#' @importFrom affiner as_coord2d degrees
 #' @importFrom rlang check_dots_empty
 NULL
 
 as_coords <- function(x, y = NULL) {
     if (is.null(y)) {
-        p <- affiner::as_coord2d(x)
+        p <- as_coord2d(x)
     } else {
         stopifnot(is.numeric(x), is.numeric(y))
-        p <- vctrs::vec_recycle_common(x = x, y = y)
+        p <- data.frame(x = x, y = y)
     }
     p
 }